refactor: 공통 레이아웃 폴더 변경
