+++ src/app/my-info/layout.tsx
@@ -1,4 +1,4 @@
-import GlobalLayout from '@/components/layout/global';
+import GlobalLayout from '@/components/common/layout';
 
 interface MyInfoPageLayoutProps {
   children: React.ReactNode;
