+++ src/app/my-library/layout.tsx
@@ -1,4 +1,4 @@
-import GlobalLayout from '@/components/layout/global';
+import GlobalLayout from '@/components/common/layout';
 
 interface MyLibraryPageLayoutProps {
   children: React.ReactNode;
