+++ src/app/search/layout.tsx
@@ -1,4 +1,4 @@
-import GlobalLayout from '@/components/layout/global';
+import GlobalLayout from '@/components/common/layout';
 
 interface SearchPageLayoutProps {
   children: React.ReactNode;
