+++ src/components/common/layout/main/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import styles from '@/components/layout/main/main.module.scss';
+import styles from '@/components/common/layout/main/main.module.scss';
 
 interface MainLayoutProps {
   children: React.ReactNode;
