+++ src/components/user/find-login-data/find-id.tsx
@@ -4,9 +4,9 @@ import { useState } from 'react';
 import { useShallow } from 'zustand/shallow';
 
 import { fetchGroo } from '@/apis';
-import BasicButton from '@/components/layout/button/basic';
-import BasicInputContainer from '@/components/layout/input/basic/container';
-import BasicInputField from '@/components/layout/input/basic/field';
+import BasicButton from '@/components/common/button/basic';
+import BasicInputContainer from '@/components/common/input/basic/container';
+import BasicInputField from '@/components/common/input/basic/field';
 import styles from '@/components/user/find-login-data/find-login-data.module.scss';
 import { useInputText } from '@/hooks/useInput';
 import useBoundStore from '@/stores';
