+++ src/components/user/login/index.tsx
@@ -5,9 +5,9 @@ import { useRouter } from 'next/navigation';
 import { useShallow } from 'zustand/shallow';
 
 import { fetchGroo } from '@/apis';
-import BasicButton from '@/components/layout/button/basic';
-import BasicInputContainer from '@/components/layout/input/basic/container';
-import BasicInputField from '@/components/layout/input/basic/field';
+import BasicButton from '@/components/common/button/basic';
+import BasicInputContainer from '@/components/common/input/basic/container';
+import BasicInputField from '@/components/common/input/basic/field';
 import styles from '@/components/user/login/login.module.scss';
 import { useInputText } from '@/hooks/useInput';
 import useBoundStore from '@/stores';
